+++ magick/annotate.c
@@ -1483,7 +1483,7 @@ static unsigned int RenderPostscript(Image *image,const DrawInfo *draw_info,
         crop_info;
 
       crop_info=GetImageBoundingBox(annotate_image,&annotate_image->exception);
-      crop_info.height=(unsigned int) ceil((resolution.y/72.0)*
+      crop_info.height=(unsigned long) ceil((resolution.y/72.0)*
         ExpandAffine(&draw_info->affine)*draw_info->pointsize-0.5);
       crop_info.y=(long) ceil((resolution.y/72.0)*extent.y/8.0-0.5);
       (void) FormatString(geometry,"%lux%lu%+ld%+ld",crop_info.width,
